[BE] Move submodule check out from setup.py
Summary: Per comment in #8203: https://github.com/pytorch/executorch/pull/8203#discussion_r1943763809

Moving the logic of checking submodule and update out from setup.py and
into install_executorch.py.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:
