+++ install_executorch.py
@@ -9,6 +9,7 @@
 import argparse
 import glob
 import itertools
+import logging
 import os
 import shutil
 import subprocess
@@ -19,6 +20,12 @@ from install_requirements import (
     python_is_compatible,
     TORCH_NIGHTLY_URL,
 )
+
+# Set up logging
+logging.basicConfig(
+    level=logging.INFO, format="%(asctime)s [ExecuTorch] %(levelname)s: %(message)s"
+)
+logger = logging.getLogger()
 
 
 def clean():
@@ -33,6 +40,83 @@ def clean():
 
 
 VALID_PYBINDS = ["coreml", "mps", "xnnpack", "training"]
+
+
+################################################################################
+# Git submodules
+################################################################################
+# The following submodules are required to be able to build ExecuTorch. If any of
+# these folders are missing or missing CMakeLists.txt, we will run
+# `git submodule update` to try to fix it. If the command fails, we will raise an
+# error.
+# An alternative to this would be to run `git submodule status` and run
+# `git submodule update` if there's any local changes. However this is a bit
+# too restrictive for users who modifies and tests the dependencies locally.
+
+# keep sorted
+REQUIRED_SUBMODULES = {
+    "ao": "LICENSE",  # No CMakeLists.txt, choose a sort of stable file to check.
+    "cpuinfo": "CMakeLists.txt",
+    "eigen": "CMakeLists.txt",
+    "flatbuffers": "CMakeLists.txt",
+    "FP16": "CMakeLists.txt",
+    "FXdiv": "CMakeLists.txt",
+    "gflags": "CMakeLists.txt",
+    "prelude": "BUCK",
+    "pthreadpool": "CMakeLists.txt",
+    "pybind11": "CMakeLists.txt",
+    "XNNPACK": "CMakeLists.txt",
+}
+
+
+def get_required_submodule_paths():
+    gitmodules_path = os.path.join(os.getcwd(), ".gitmodules")
+
+    if not os.path.isfile(gitmodules_path):
+        logger.error(".gitmodules file not found.")
+        exit(1)
+
+    with open(gitmodules_path, "r") as file:
+        lines = file.readlines()
+
+    # Extract paths of required submodules
+    required_paths = {}
+    for line in lines:
+        if line.strip().startswith("path ="):
+            path = line.split("=")[1].strip()
+            for submodule, file_name in REQUIRED_SUBMODULES.items():
+                if submodule in path:
+                    required_paths[path] = file_name
+    return required_paths
+
+
+def check_and_update_submodules():
+    def check_folder(folder: str, file: str) -> bool:
+        return os.path.isdir(folder) and os.path.isfile(os.path.join(folder, file))
+
+    # Check if the directories exist for each required submodule
+    missing_submodules = {}
+    for path, file in get_required_submodule_paths().items():
+        if not check_folder(path, file):
+            missing_submodules[path] = file
+
+    # If any required submodule directories are missing, update them
+    if missing_submodules:
+        logger.warning("Some required submodules are missing. Updating submodules...")
+        try:
+            subprocess.check_call(["git", "submodule", "sync"])
+            subprocess.check_call(["git", "submodule", "update", "--init"])
+        except subprocess.CalledProcessError as e:
+            logger.error(f"Error updating submodules: {e}")
+            exit(1)
+
+        # After updating submodules, check again
+        for path, file in missing_submodules.items():
+            if not check_folder(path, file):
+                logger.error(f"{file} not found in {path}.")
+                logger.error("Please run `git submodule update --init`.")
+                exit(1)
+    logger.info("All required submodules are present.")
 
 
 def main(args):
@@ -119,6 +203,9 @@ def main(args):
     # Set environment variables
     os.environ["EXECUTORCH_BUILD_PYBIND"] = EXECUTORCH_BUILD_PYBIND
     os.environ["CMAKE_ARGS"] = CMAKE_ARGS
+
+    # Check if the required submodules are present and update them if not
+    check_and_update_submodules()
 
     # Run the pip install command
     subprocess.run(
